+++ source/Core/Module/ModuleInstaller.php
@@ -407,9 +407,11 @@ class ModuleInstaller extends \OxidEsales\Eshop\Core\Base
     }
 
     /**
-     * Add module templates to database.
-     *
-     * @todo extract oxtplblocks query to ]
+     * Set module templates in the database.
+     * we do not use delete and add combination because
+     * the combination of deleting and adding does unnecessary writes and so it does not scale
+     * also it's more likely to get race conditions (in the moment the blocks are deleted)
+     * @todo extract oxtplblocks query to ModuleTemplateBlockRepository
      *
      * @param array  $moduleBlocks Module blocks array
      * @param string $moduleId     Module id
@@ -418,10 +420,11 @@ class ModuleInstaller extends \OxidEsales\Eshop\Core\Base
     {
         $shopId = $this->getConfig()->getShopId();
         $db = \OxidEsales\Eshop\Core\DatabaseProvider::getDb();
-
         if (is_array($moduleBlocks)) {
+            $knownBlocks = ['dummy'];
             foreach ($moduleBlocks as $moduleBlock) {
-                $id = Registry::getUtilsObject()->generateUId();
+                $id = md5($moduleId . json_encode($moduleBlock) . $shopId);
+                $knownBlocks[] = $id;
 
                 $template = $moduleBlock["template"];
                 $position = isset($moduleBlock['position']) && is_numeric($moduleBlock['position']) ?
@@ -433,7 +436,18 @@ class ModuleInstaller extends \OxidEsales\Eshop\Core\Base
                 $theme = isset($moduleBlock['theme']) ? $moduleBlock['theme'] : '';
 
                 $sql = "INSERT INTO `oxtplblocks` (`OXID`, `OXACTIVE`, `OXSHOPID`, `OXTHEME`, `OXTEMPLATE`, `OXBLOCKNAME`, `OXPOS`, `OXFILE`, `OXMODULE`)
-                         VALUES (?, 1, ?, ?, ?, ?, ?, ?, ?)";
+                     VALUES (?, 1, ?, ?, ?, ?, ?, ?, ?)
+                     ON DUPLICATE KEY UPDATE
+                      `OXID` = VALUES(OXID),
+                      `OXACTIVE` = VALUES(OXACTIVE),
+                      `OXSHOPID` = VALUES(OXSHOPID),
+                      `OXTHEME`  = VALUES(OXTHEME),
+                      `OXTEMPLATE` = VALUES(OXTEMPLATE),
+                      `OXBLOCKNAME` = VALUES(OXBLOCKNAME),
+                      `OXPOS` = VALUES(OXPOS),
+                      `OXFILE` = VALUES(OXFILE),
+                      `OXMODULE` = VALUES(OXMODULE)
+                     ";
 
                 $db->execute(
                     $sql,
@@ -449,6 +463,12 @@ class ModuleInstaller extends \OxidEsales\Eshop\Core\Base
                     ]
                 );
             }
+
+            $commaSeparatedListOfKnownBlocks = join(',', $db->quoteArray($knownBlocks));
+            $db->execute(
+                "DELETE FROM oxtplblocks WHERE OXSHOPID = ? AND OXMODULE = ? AND OXID NOT IN ($commaSeparatedListOfKnownBlocks) ",
+                array($shopId, $moduleId)
+            );
         }
     }